refactor the codes
